Refactored backend directory #255
- Split the tests into different suites
- Created new module just for the signals
- Fixed linting errors accross all files
